% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{a_step}
\alias{a_step}
\title{Minimization over scores in Bernoulli PCA}
\usage{
a_step(X, a_cur, v_cur)
}
\arguments{
\item{X}{The n x p binary matrix with samples along rows which we want to
decompose using exponential family PCA.}

\item{a_cur}{The current values of the latent scores. Used to initialize the
search for the next iteration's scores.}

\item{v_cur}{The current values of the loadings.}
}
\value{
The scores at the next iteration.
}
\description{
Minimization over scores in Bernoulli PCA
}

