% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bern_breg_grad_a}
\alias{bern_breg_grad_a}
\title{Gradient of Bregman Divergence with respect to parameter a}
\usage{
bern_breg_grad_a(x, v)
}
\arguments{
\item{x}{A vector representing a single sample from the matrix we are
decomposing with PCA.}

\item{v}{The current value of the loadings to keep fixed during the
optimization over scores.}
}
\value{
Gradient of the function to optimize over a, corresponding to the
current values of the loadings and the i^th sample of X.
}
\description{
Gradient of Bregman Divergence with respect to parameter a
}

