% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bern_breg_grad_v}
\alias{bern_breg_grad_v}
\title{Gradient of Bregman Divergence with respect to parameter v}
\usage{
bern_breg_grad_v(x, a)
}
\arguments{
\item{x}{A vector representing a single dimension from the matrix we are
decomposing with PCA.}

\item{a}{The current value of the scores to keep fixed during the
optimization over loadings.}
}
\value{
The gradient of the function to optimize over v, corresponding to the
current values of the scores and the i^th dimension of X.
}
\description{
Gradient of Bregman Divergence with respect to parameter v
}

