% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bern_breg_obj_a}
\alias{bern_breg_obj_a}
\title{Bernoulli Bregman Divergence Objective as a Function of parameter a}
\usage{
bern_breg_obj_a(x, v)
}
\arguments{
\item{x}{A vector representing a single sample from the matrix we are
decomposing with PCA.}

\item{v}{The current value of the loadings to keep fixed during the
optimization over scores.}
}
\value{
The function to optimize over a, corresponding to the current values
of the loadings and the i^th sample of X.
}
\description{
To fit Exponential Family PCA, we need to alternate
n + d optimizations, n with respect to the scores and d with respect to the
loadings. This function helps return the function to optimize in each of the
n optimizations across scores.
}

