% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bern_breg_obj_v}
\alias{bern_breg_obj_v}
\title{Bernoulli Bregman Divergence Objective as a Function of parameter v}
\usage{
bern_breg_obj_v(x, a)
}
\arguments{
\item{x}{A vector representing a single dimension from the matrix we are
decomposing with PCA.}

\item{a}{The current value of the scores to keep fixed during the
optimization over loadings.}
}
\value{
The function to optimize over v, corresponding to the current values
of the scores and the i^th dimension of X.
}
\description{
To fit Exponential Family PCA, we need to alternate
n + d optimizations, n with respect to the scores and d with respect to the
loadings. This function helps return the function to optimize in each of the
d optimizations across loadings coordinates.
}

