% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bern_exp_pca}
\alias{bern_exp_pca}
\title{Bernoulli PCA with one component}
\usage{
bern_exp_pca(X, a0 = NULL, v0 = NULL, iter_max = 100, eps = 1e-04)
}
\arguments{
\item{X}{The n x p binary matrix with samples along rows which we want to
decompose using exponential family PCA.}

\item{a0}{Initial values for the latent scores. If not supplied, initialized
with standard normals.}

\item{v0}{Initial values for loading vector. If not supplied, initialized
with standard normals.}

\item{iter_max}{The maximum number of iterations to run the PCA.}

\item{eps}{The convergence criterion. If the mean change in the scores is
less than eps, we return.}
}
\value{
A list containing the converged values of a and v, along with traces
of their values across the optimization [A and V, the columns are values
across iterations].
}
\description{
Bernoulli PCA with one component
}
\references{
Collins, Michael, Sanjoy Dasgupta, and Robert E. Schapire. "A generalization of principal components analysis to the exponential family." Advances in neural information processing systems. 2001.
}

