% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bern_grad_a}
\alias{bern_grad_a}
\title{Gradient of Bregman Loss in Bernoulli Case with respect to a}
\usage{
bern_grad_a(x, a, v, s, lambda, mu0)
}
\description{
Gradient of Bregman Loss in Bernoulli Case with respect to a
}

