% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exp_pca.R
\name{exp_pca}
\alias{exp_pca}
\title{Exponential Family PCA via Alternating Minimization}
\usage{
exp_pca(obj_fun, obj_grad_a, obj_grad_v, X, n_comp, n_cycle, n_iter, eps,
  lambda, mu0, verbose = FALSE)
}
\arguments{
\item{obj_fun}{The main bregman loss function to attempt to minimize. See
bern_breg() for an example of the required input / output format.}

\item{obj_grad_a}{The derivative of the bregman loss with respect to the
scores. See bern_brad_grad_a() for an example of the required input / output
format.}

\item{obj_grad_v}{The derivative of the bregman loss with respect to the
loadings. See bern_brad_grad_a() for an example of the required input /
output format.}

\item{X}{The n x p binary matrix with samples along rows which we want to
decompose using exponential family PCA.}

\item{n_comp}{How many principal components should we return?}

\item{n_cycle}{How many times should the iterative optimization pass through
across each component?}

\item{n_iter}{The maximum number of iterations to run the PCA.}

\item{eps}{The convergence criterion. If the mean change in the scores is
less than eps, we return.}

\item{lambda}{The regularization parameter in the optimization.}

\item{mu0}{The value to regularize towards.}

\item{verbose}{Print iterations of procedure?}
}
\value{
A list containing the converged values of A and V.
}
\description{
General exponential-family PCA algorithm, within which different
family-specific bregman losses can be input.
}
\references{
Collins, Michael, Sanjoy Dasgupta, and Robert E. Schapire. "A generalization of principal components analysis to the exponential family." Advances in neural information processing systems. 2001.
}

