% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{generate_bern_data}
\alias{generate_bern_data}
\title{Simulate test data for bernoulli PCA}
\usage{
generate_bern_data(k = 4, n = 100, d = 10, p = 0.5, eps = 0.05)
}
\arguments{
\item{k}{The number of latent binary classes.}

\item{n}{The number of samples to generate.}

\item{d}{The dimension of the samples.}

\item{p}{The probability of a 1 in each coordinate of each latent class.}

\item{eps}{The probability of flipping the sign in the observed data.}
}
\description{
Simulate test data for bernoulli PCA
}

