% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_exp_pca}
\alias{poisson_exp_pca}
\title{Poisson PCA}
\usage{
poisson_exp_pca(X, n_comp = 2, n_cycle = 30, n_iter = 30, eps = 1e-04,
  lambda = 0.001, mu0 = NULL)
}
\arguments{
\item{X}{The n x p Poisson matrix with samples along rows which we want to
decompose using exponential family PCA.}

\item{n_comp}{How many principal components should we return?}

\item{n_cycle}{How many times should the iterative optimization pass through
across each component?}

\item{n_iter}{The maximum number of iterations to run the PCA.}

\item{eps}{The convergence criterion. If the mean change in the scores is
less than eps, we return.}

\item{lambda}{The regularization parameter in the optimization.}

\item{mu0}{The value to regularize towards.}
}
\value{
A list containing the converged values of A and V.
}
\description{
Poisson PCA
}
\references{
Collins, Michael, Sanjoy Dasgupta, and Robert E. Schapire. "A generalization of principal components analysis to the exponential family." Advances in neural information processing systems. 2001.
}

