% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_grad_a}
\alias{poisson_grad_a}
\title{Gradient of Bregman Loss for Poisson data with respect a}
\usage{
poisson_grad_a(x, a, v, s, lambda, mu0)
}
\arguments{
\item{x}{A single row of column of x, for which we want to minimize distance
to the mean.}

\item{a}{Either a vector or scale representing the latent scores.}

\item{v}{Either a vector or scale representing the latent loadings.}

\item{s}{Either the s^th row or column of the matrix AV^T - a_{c}v_{c}^{T}}

\item{lambda}{The regularization parameter in the optimization.}

\item{mu0}{The value to regularize towards.}
}
\value{
The value of the gradient of the bregman loss with respect to the
parameter v.
}
\description{
Gradient of Bregman Loss for Poisson data with respect a
}

