% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{r}
\alias{r}
\title{Logistic Function}
\usage{
r(x)
}
\arguments{
\item{x}{Input vector or scale for logistic fun.}
}
\value{
v Value of logistic function at input.
}
\description{
Logistic Function
}

