% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exp_pca_updates.R
\name{update_a_ic}
\alias{update_a_ic}
\title{Update one element of scores A}
\usage{
update_a_ic(obj_fun, obj_grad_a, x_i, a_ic, v_c, s_i, lambda, mu0)
}
\arguments{
\item{obj_fun}{Version of loss given by bregman divergence between data and
link to use in current updates.}

\item{obj_grad_a}{Gradient of the provided loss, with respect to a.}

\item{x_i}{The i^th sample of X.}

\item{a_ic}{The initial value of a_ic to start the search from.}

\item{v_c}{The c^th component of V.}

\item{s_i}{The s^th row of the matrix AV^T - a_{c}v_{c}^{T}}

\item{lambda}{The regularization parameter in the optimizatyion.}

\item{mu0}{The value to regularize towards.}
}
\value{
a_ic The updated version of a_ic.
}
\description{
Update one element of scores A
}

