% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exp_pca_updates.R
\name{update_v_jc}
\alias{update_v_jc}
\title{Update one element of loadings V}
\usage{
update_v_jc(obj_fun, obj_grad_v, x_j, v_jc, a_c, s_j, lambda, mu0)
}
\arguments{
\item{obj_fun}{Version of loss given by bregman divergence between data and
link to use in current updates.}

\item{obj_grad_v}{Gradient of the provided loss, with respect to v.}

\item{x_j}{The j^th dimension of X.}

\item{v_jc}{The initial value of v_jc to start the search from.}

\item{a_c}{The c^th component of A.}

\item{s_j}{The s^th dimension of the matrix AV^T - a_{c}v_{c}^{T}.}

\item{lambda}{The regularization parameter in the optimization.}

\item{mu0}{The value to regularize towards.}
}
\value{
v_jc The updated version of v_jc.
}
\description{
Update one element of loadings V
}

