% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{v_step}
\alias{v_step}
\title{Minimization over loadings in Bernoulli PCA}
\usage{
v_step(X, a_cur, v_cur)
}
\arguments{
\item{X}{The n x p binary matrix with samples along rows which we want to
decompose using exponential family PCA.}

\item{a_cur}{The current values of the latent scores.}

\item{v_cur}{The current values of the loadings. Used to initialize the
search for the next iteration's loadings.}
}
\value{
The loadings at the next iteration.
}
\description{
Minimization over loadings in Bernoulli PCA
}

