% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exp_pca_updates.R
\name{exp_pca}
\alias{exp_pca}
\title{Exponential Family PCA via Alternating Minimization}
\usage{
exp_pca(obj_fun, obj_grad_a, obj_grad_v, X, n_comp, n_cycle, n_iter, eps,
  lambda, mu0)
}
\description{
Exponential Family PCA via Alternating Minimization
}

